+++ src/main/kotlin/io/snyk/plugin/cli/ConsoleCommandRunner.kt
@@ -10,6 +10,7 @@ import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Key
+import com.intellij.util.net.HttpConfigurable
 import io.snyk.plugin.controlExternalProcessWithProgressIndicator
 import io.snyk.plugin.getWaitForResultsTimeout
 import io.snyk.plugin.pluginSettings
@@ -98,6 +99,18 @@ open class ConsoleCommandRunner {
         commandLine.environment["SNYK_INTEGRATION_VERSION"] = pluginInfo.integrationVersion
         commandLine.environment["SNYK_INTEGRATION_ENVIRONMENT"] = pluginInfo.integrationEnvironment
         commandLine.environment["SNYK_INTEGRATION_ENVIRONMENT_VERSION"] = pluginInfo.integrationEnvironmentVersion
+        val proxySettings = HttpConfigurable.getInstance()
+        val proxyHost = proxySettings.PROXY_HOST
+        if (proxySettings != null && proxySettings.USE_HTTP_PROXY && proxyHost.isNotEmpty()) {
+            var authentication = ""
+            if (proxySettings.PROXY_AUTHENTICATION) {
+                val auth =
+                    proxySettings.getPromptedAuthentication(proxyHost, "Snyk: Please enter your proxy password")
+                authentication = auth.userName + ":" + String(auth.password) + "@"
+            }
+            commandLine.environment["http_proxy"] = "http://$authentication$proxyHost:${proxySettings.PROXY_PORT}"
+            commandLine.environment["https_proxy"] = "http://$authentication$proxyHost:${proxySettings.PROXY_PORT}"
+        }
     }
 
     companion object {